move `copies.rs` to `ifs/mod.rs`
all the lints work with `if`-expressions
